chore: 프로젝트 디테일 컴포넌트 재사용
